+++ ergo_runtime/src/traits.rs
@@ -8,7 +8,7 @@ mod stored;
 mod type_name;
 mod value_by_content;
 
-pub use bind::{bind, delay_bind, Bind, BindImpl};
+pub use bind::{bind, bind_error, delay_bind, Bind, BindImpl};
 pub use display::{display, Display, DisplayImpl};
 pub use into::{into, into_sourced, IntoTyped, IntoTypedImpl};
 pub use nested::{force_value_nested, Nested, NestedImpl};